Go back to linear inheritence for P0848.
Linear inheritence from a base implementation seems to be necessary for this
task, as that's the only way I can seem to allow a trait's policy implementation
to override the base's default implementation.

Essentially in a situtation like below we need to avoid ever calling `A`'s
default copy constructor, because `B` provides the only functionality to run.

```
struct A {};
struct add_copy_ctor { B(const B&) {} };
struct C : A, add_copy_ctor {};
```