Unconditionally add the monotype to the storage union.
Previous testing only showed that certain versions of clang wanted the monotype,
but some real world cases showed gcc wanting it as well in some contexts. This
commit just unconditionally adds the monotype to the storage union. At some
point in the future this shouldn't be necessary.

This also extends the unit tests with tests to check for this regression, and
switches from `mutable` to `const_cast` in the runtime tests `Foo` struct to
avoid problems related to mutable fields in `constexpr`.