Remove Catch2 dependency, and expand tests.
As a constexpr-based header library, Catch2 isn't really ideal as a unit test
platform. Test failures have terrible symbols injected, and it adds a bunch of
extra compile time. This commit removes that dependency.

We also want to try and ensure that widest test coverage that we can, which
means hosting the vector template with all possible combinations of trivial
traits. This commit adds a simple `common.decl` file that uses normal macro
magic to expand statements for every possible combination of the 5 trivial
traits that we care about. It uses this functionality to create a test-wide
family of Foo classes, and then instantiates existing cvector tests with that
template.

Finally we switch to using a builtin test CHECK macro instead of `assert` in
order to get access to customizable failure logic.